Create projects, use Moq for simple test of user settings
The class under test shows the use of application settings, which are stored
in a "user" file. This means that to unit test the class, a test version of
this file should be used. Doing this literally would likely be messy. Instead,
use a mock of the settings in the unit tests. I have decided to try the Moq
framework because it seems popular.

This initial feature "run count" is not very interesting but hopefully it
shows how Moq can be used.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>